verify: fix FUN_000a1510 delinked alignment (41%→86%), add @reg for game_allegiance_set
- Re-export delinked/particles.obj after creating FUN_000a1590 boundary in Ghidra
  (hidden 204-byte function was merged into FUN_000a1510's reference, inflating it
  from 45 to 144 instructions). FUN_000a1510 now matches 85.7% (was 41.1%).
- Fix FUN_000a1510 param type: int -> int16_t (MOVSX in disasm confirms short).
- Add @eax/@ebx register annotations to game_allegiance_set (entry in EAX,
  friendship in BL, force on stack). All callers are ported so no runtime
  impact, but fixes ABI metadata and explains 0/50 equiv false-failure.
- Add FUN_000a1590 to kb.json as unported particles.obj function.

+++ src/halo/effects/particles.c
@@ -152,7 +152,7 @@ void particle_delete(int datum_handle)
 
 /* Delete all particles owned by a local player that have an attached
    object (flag 0x40 set and object handle != -1). */
-void FUN_000a1510(int local_player_index)
+void FUN_000a1510(int16_t local_player_index)
 {
   int handle;
   char *datum;
